+++ CHANGELOG.md
@@ -607,6 +607,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 - Use 2 bit record the number of occurrences of 1: `00`, `01`, `10`
 - Use **Karnaugh map** to facilitate the simplification of Boolean algebra
 
+## 0141 - Linked List Cycle
+
+- Two points: slow and fast
+
 ## 0415 - Add strings
 
 - Full adder